requirements.txt
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
pyyaml>=6.0
requests>=2.28.0
matplotlib>=3.7.0
seaborn>=0.12.0
pytest>=7.3.0
tabulate>=0.9.0
